requirements.txt
git+https://github.com/Pycord-Development/pycord
replit
flask
python-dotenv
statuspageio